# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"BasicResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"description": "Basic JSON response for controllers"
			},
			"GoodbyeResponse": {
				"properties": {
					"date": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"date",
					"message"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "code-verifier-backend",
		"version": "1.0.0",
		"description": "Backend Node + Express + TS + Swagger",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@chapa0723"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/hello": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Promise of BasicResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponse"
								}
							}
						}
					}
				},
				"description": "EndPoint to retreive a Massage \"Hello {name}\" in JSON",
				"tags": [
					"HelloController"
				],
				"security": [],
				"parameters": [
					{
						"description": "of user to be greeted",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/goodbye": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Promise of BasicResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GoodbyeResponse"
								}
							}
						}
					}
				},
				"description": "EndPoint to retreive a Massage \"Hello {name}\" in JSON",
				"tags": [
					"GoodbyeController"
				],
				"security": [],
				"parameters": [
					{
						"description": "of user to be greeted",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to retreive all users in the collection \"Users\" of DB",
				"tags": [
					"UsersController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}